Split driver tests by state machine
